src/main/java/com/yelp/clientlib/entities/Region.java
package com.yelp.clientlib.entities;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.databind.PropertyNamingStrategy;
import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
import com.google.auto.value.AutoValue;

import java.io.Serializable;

@AutoValue
@JsonIgnoreProperties(ignoreUnknown = true)
@JsonDeserialize(builder = AutoValue_Region.Builder.class)
public abstract class Region implements Serializable {

    public abstract Coordinate center();

    public abstract Span span();

    @AutoValue.Builder
    @JsonPOJOBuilder(withPrefix = "")
    @JsonNaming(PropertyNamingStrategy.LowerCaseWithUnderscoresStrategy.class)
    public abstract static class Builder {

        public abstract Builder center(Coordinate center);

        public abstract Builder span(Span span);

        public abstract Region build();
    }

    public static Builder builder() {
        return new AutoValue_Region.Builder();
    }
}
